IS-36 refactor of package name
