--- idp/src/main/java/se/elegnamnden/eid/idp/authn/model/SignMessageModel.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2016-2018 Litsec AB
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package se.elegnamnden.eid.idp.authn.model;
-
-import lombok.Data;
-
-/**
- * Model object for a SignMessage.
- * 
- * @author Martin Lindström (martin.lindstrom@litsec.se)
- */
-@Data
-public class SignMessageModel {
-  
-  public enum DisplayType { MONOSPACE_TEXT, HTML }; 
-  
-  /** The type of sign message. */
-  private DisplayType displayType;
-  
-  /** The HTML. */
-  private String html;
-
-}
