--- idp/src/main/java/se/elegnamnden/eid/idp/authn/model/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Domain objects used for the simulated authentication process. 
- */
-package se.elegnamnden.eid.idp.authn.model;
