--- idp/src/main/java/se/elegnamnden/eid/idp/authn/service/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Authentication service implementations. 
- */
-package se.elegnamnden.eid.idp.authn.service;
