+++ idp/src/main/java/se/swedenconnect/eid/idp/authn/model/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Domain objects used for the simulated authentication process. 
+ */
+package se.swedenconnect.eid.idp.authn.model;
