+++ idp/src/main/java/se/swedenconnect/eid/idp/authn/service/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Authentication service implementations. 
+ */
+package se.swedenconnect.eid.idp.authn.service;
