+++ idp/src/main/shibboleth/config/conf/logback.xml
@@ -52,7 +52,8 @@
     <!-- ELN & eIDAS -->
     <logger name="se.litsec.shibboleth.idp" level="${idp.loglevel.eidas:-INFO}" />
     <logger name="se.litsec" level="${idp.loglevel.eln:-INFO}" />
-    <logger name="se.elegnamnden" level="${idp.loglevel.eln:-INFO}" />    
+    <logger name="se.elegnamnden" level="${idp.loglevel.eln:-INFO}" />
+    <logger name="se.swedenconnect" level="${idp.loglevel.eln:-INFO}" />    
 
     <!-- Especially chatty. -->
     <logger name="se.litsec.shibboleth.idp.attribute.resolver.SAML2AttributeNameToIdMapperService" level="INFO" />
